Implement explicit and implicit casting validation with comprehensive checks and error reporting; add tests for casting scenarios
