+++ compiler/internal/semantic/typecheck/expressions.go
@@ -1,8 +1,6 @@
 package typecheck
 
 import (
-	"errors"
-	"ferret/compiler/colors"
 	"ferret/compiler/internal/frontend/ast"
 	"ferret/compiler/internal/modules"
 	"ferret/compiler/internal/report"
@@ -12,7 +10,6 @@ import (
 	"ferret/compiler/internal/source"
 	"ferret/compiler/internal/types"
 	"fmt"
-	"strings"
 )
 
 // evaluateExpressionType infers the semantic type from an AST expression
@@ -133,130 +130,6 @@ func checkFunctionCallType(r *analyzer.AnalyzerNode, call *ast.FunctionCallExpr,
 	return funcType.ReturnType
 }
 
-// checkCastExprType validates type cast expressions and returns the target type
-func checkCastExprType(r *analyzer.AnalyzerNode, cast *ast.CastExpr, cm *modules.Module) stype.Type {
-	// Evaluate the source expression type
-	sourceType := evaluateExpressionType(r, *cast.Value, cm)
-	if sourceType == nil {
-		return nil
-	}
-
-	// Convert AST target type to semantic type
-	targetType, err := semantic.DeriveSemanticType(cast.TargetType, cm)
-	if err != nil || targetType == nil {
-		r.Ctx.Reports.AddSemanticError(
-			r.Program.FullPath,
-			cast.Loc(),
-			fmt.Sprintf("invalid target type in cast expression: %v", err),
-			report.TYPECHECK_PHASE,
-		)
-		return nil
-	}
-
-	// Check if the cast is valid
-	isValid, err := isExplicitCastable(sourceType, targetType)
-	if err != nil || !isValid {
-		r.Ctx.Reports.AddSemanticError(
-			r.Program.FullPath,
-			cast.Loc(),
-			fmt.Sprintf("cannot cast from '%s' to '%s': %v", sourceType.String(), targetType.String(), err),
-			report.TYPECHECK_PHASE,
-		)
-		return nil // Return nil if cast is invalid
-	}
-
-	return targetType
-}
-
-func isExplicitCastable(sourceType, targetType stype.Type) (bool, error) {
-	// primitive types can be casted to each other
-	if sourceType == nil || targetType == nil {
-		return false, errors.New("source or target type is nil")
-	}
-	
-	if isImplicitCastable(sourceType, targetType) {
-		return true, nil // Implicit cast is also valid for explicit cast
-	}
-
-	sourceUnwrapped := semantic.UnwrapType(sourceType) // Unwrap any type aliases
-	targetUnwrapped := semantic.UnwrapType(targetType) // Unwrap any type aliases
-
-	// Check if both are primitive types
-	if _, ok := sourceUnwrapped.(*stype.PrimitiveType); ok {
-		if _, ok := targetUnwrapped.(*stype.PrimitiveType); ok {
-			// Both are primitive types, check if they are castable
-			return isPrimitiveExplicitCastable(sourceUnwrapped, targetUnwrapped)
-		}
-	}
-
-	//structs can be casted to other struct and interfaces and interfaces can be casted to structs
-	if ss, ok := sourceUnwrapped.(*stype.StructType); ok {
-		//target must be a struct or interface. for now skip interface
-		if ts, ok := targetUnwrapped.(*stype.StructType); ok {
-			//both are struct
-			return isStructExplicitCastable(ss, ts)
-		}
-	}
-
-	return false, fmt.Errorf("no valid cast found from '%s' to '%s'", sourceType.String(), targetType.String())
-}
-
-func isStructExplicitCastable(sourceType, targetType *stype.StructType) (bool, error) {
-	//targets all properties must present
-
-	fieldErrors := make([]string, 0, len(targetType.Fields))
-
-	for fieldName, fieldType := range targetType.Fields {
-		if sourceFieldType, exists := sourceType.Fields[fieldName]; !exists {
-			fieldErrors = append(fieldErrors, colors.RED.Sprintf(" - missing field '%s'", fieldName))
-		} else if !isImplicitCastable(sourceFieldType, fieldType) {
-			fieldErrors = append(fieldErrors, colors.PURPLE.Sprintf(" - field '%s' type required '%s', but got '%s'", fieldName, fieldType.String(), sourceFieldType.String()))
-		}
-	}
-
-	if len(fieldErrors) > 0 {
-		errMsg := colors.WHITE.Sprintf("\n%s and %s has field missmatch\n", sourceType.String(), targetType.String())
-		return false, fmt.Errorf("%s%s", errMsg, strings.Join(fieldErrors, "\n"))
-	}
-
-	fmt.Printf("Successfully casted struct '%s' to '%s'\n", sourceType.String(), targetType.String())
-
-	return true, nil
-}
-
-func isPrimitiveExplicitCastable(sourceType, targetType stype.Type) (bool, error) {
-	sourcePrim, sOk := sourceType.(*stype.PrimitiveType)
-	targetPrim, tOk := targetType.(*stype.PrimitiveType)
-
-	if !sOk || !tOk {
-		return false, errors.New("both source and target must be primitive types")
-	}
-
-	// Allow ALL numeric to numeric casting with explicit "as" keyword
-	// The developer explicitly requests the conversion, so allow both widening and narrowing
-	if types.IsNumericTypeName(sourcePrim.Name) && types.IsNumericTypeName(targetPrim.Name) {
-		return true, nil
-	}
-
-	// Special case: byte can be cast to/from u8 and i8
-	if sourcePrim.Name == types.BYTE {
-		res := targetPrim.Name == types.UINT8 || targetPrim.Name == types.INT8
-		if res {
-			return res, nil
-		}
-		return false, fmt.Errorf("byte and %s are incompatible types", targetPrim.Name)
-	}
-	if targetPrim.Name == types.BYTE {
-		res := sourcePrim.Name == types.UINT8 || sourcePrim.Name == types.INT8
-		if res {
-			return res, nil
-		}
-		return false, fmt.Errorf("byte and %s are incompatible types", sourcePrim.Name)
-	}
-
-	// No valid cast found
-	return false, fmt.Errorf("no valid cast found from '%s' to '%s'", sourceType.String(), targetType.String())
-}
 
 // checkFieldAccessType handles struct field access and method access
 func checkFieldAccessType(r *analyzer.AnalyzerNode, fieldAccess *ast.FieldAccessExpr, cm *modules.Module) stype.Type {
