+++ compiler/internal/semantic/typecheck/utils.go
@@ -1,7 +1,6 @@
 package typecheck
 
 import (
-	"ferret/compiler/colors"
 	"ferret/compiler/internal/frontend/ast"
 	"ferret/compiler/internal/modules"
 	"ferret/compiler/internal/report"
@@ -11,135 +10,6 @@ import (
 	"ferret/compiler/internal/types"
 	"fmt"
 )
-
-// ===== CORE ASSIGNABILITY CHECK =====
-
-// isImplicitCastable checks if a value of type 'source' can be assigned to 'target'
-// Note: This function has limited type resolution capability. For full alias resolution,
-// the type checker should use resolveTypeAlias with analyzer context.
-func isImplicitCastable(target, source stype.Type) bool {
-	// Handle user types (aliases) - limited resolution without symbol table access
-	resolvedTarget := semantic.UnwrapType(target)
-	resolvedSource := semantic.UnwrapType(source)
-
-	colors.PURPLE.Printf("Checking assignability: %v → %v ", resolvedSource, resolvedTarget)
-
-	if isPrimitiveImplicitCastable(resolvedTarget, resolvedSource) || isArrayImplicitCastable(resolvedTarget, resolvedSource) || isFunctionImplicitCastable(resolvedTarget, resolvedSource) || isStructImpliticCastable(resolvedTarget, resolvedSource) {
-		colors.GREEN.Println(" ✔ ")
-		return true
-	}
-
-	colors.RED.Println(" ✘ ")
-
-	return false
-}
-
-// ===== HELPER FUNCTIONS =====
-
-// isPrimitiveImplicitCastable checks if source can be promoted to target (implicit conversion)
-func isPrimitiveImplicitCastable(target, source stype.Type) bool {
-	targetPrim, targetOk := target.(*stype.PrimitiveType)
-	sourcePrim, sourceOk := source.(*stype.PrimitiveType)
-
-	if !targetOk || !sourceOk {
-		return false
-	}
-
-	if targetPrim.TypeName() == sourcePrim.TypeName() {
-		return true // Same type is always assignable
-	}
-
-	// Define promotion rules
-	promotions := map[types.TYPE_NAME][]types.TYPE_NAME{
-		// Integer promotions (smaller -> larger)
-		types.INT16:  {types.INT8},
-		types.INT32:  {types.INT8, types.INT16},
-		types.INT64:  {types.INT8, types.INT16, types.INT32},
-		types.UINT16: {types.UINT8, types.BYTE},
-		types.UINT32: {types.UINT8, types.UINT16, types.BYTE},
-		types.UINT64: {types.UINT8, types.UINT16, types.UINT32, types.BYTE},
-
-		// Float promotions (int -> float, smaller float -> larger float)
-		types.FLOAT32: {types.INT8, types.INT16, types.UINT8, types.UINT16, types.BYTE},
-		types.FLOAT64: {types.INT8, types.INT16, types.INT32, types.UINT8, types.UINT16, types.UINT32, types.BYTE, types.FLOAT32},
-	}
-
-	if allowedSources, exists := promotions[targetPrim.Name]; exists {
-		for _, allowedSource := range allowedSources {
-			if sourcePrim.Name == allowedSource {
-				return true
-			}
-		}
-	}
-
-	return false
-}
-
-// isArrayImplicitCastable checks array type compatibility
-func isArrayImplicitCastable(target, source stype.Type) bool {
-	targetArray, targetOk := target.(*stype.ArrayType)
-	sourceArray, sourceOk := source.(*stype.ArrayType)
-
-	if !targetOk || !sourceOk {
-		return false
-	}
-
-	return isImplicitCastable(targetArray.ElementType, sourceArray.ElementType)
-}
-
-// isFunctionImplicitCastable checks function type compatibility
-func isFunctionImplicitCastable(target, source stype.Type) bool {
-	targetFunc, targetOk := target.(*stype.FunctionType)
-	sourceFunc, sourceOk := source.(*stype.FunctionType)
-
-	if !targetOk || !sourceOk {
-		return false
-	}
-
-	// Parameter and return type counts must match
-	if len(targetFunc.Parameters) != len(sourceFunc.Parameters) {
-		return false
-	}
-
-	// Parameters are contravariant, returns are covariant
-	for i := range targetFunc.Parameters {
-		if !isImplicitCastable(sourceFunc.Parameters[i], targetFunc.Parameters[i]) {
-			return false
-		}
-	}
-
-	// Compare return types
-	if targetFunc.ReturnType == nil && sourceFunc.ReturnType == nil {
-		return true
-	}
-	if targetFunc.ReturnType == nil || sourceFunc.ReturnType == nil {
-		return false
-	}
-
-	return isImplicitCastable(targetFunc.ReturnType, sourceFunc.ReturnType)
-}
-
-// isStructImpliticCastable checks structural compatibility of structs
-func isStructImpliticCastable(target, source stype.Type) bool {
-	targetStruct, targetOk := target.(*stype.StructType)
-	sourceStruct, sourceOk := source.(*stype.StructType)
-
-	if !targetOk || !sourceOk {
-		return false
-	}
-
-	// Source must have all fields that target requires, with compatible types
-	for fieldName, targetFieldType := range targetStruct.Fields {
-		sourceFieldType, exists := sourceStruct.Fields[fieldName]
-		if !exists || !isImplicitCastable(targetFieldType, sourceFieldType) {
-			return false
-		}
-	}
-
-	return true
-}
-
-// ===== EXPRESSION TYPE INFERENCE HELPERS =====
 
 // checkIdentifierType gets the type of an identifier from the symbol table
 func checkIdentifierType(e *ast.IdentifierExpr, cm *modules.Module) stype.Type {
